# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitForge Workout App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name?: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface UpdateMeInput {
  name: string;
}

export interface Exercise {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  muscleGroup: string;
  equipment: string;
  /** @nullable */
  instructions?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isCustom: boolean;
  createdAt: string;
}

export interface ExerciseInput {
  name: string;
  description?: string;
  category: string;
  muscleGroup: string;
  equipment: string;
  instructions?: string;
}

export interface Workout {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  difficulty: string;
  estimatedMinutes: number;
  /** @nullable */
  category?: string | null;
  isAiGenerated?: boolean;
  exerciseCount?: number;
  createdAt: string;
}

export interface WorkoutExerciseInput {
  exerciseId: number;
  sets: number;
  reps: number;
  /** @nullable */
  weight?: number | null;
  restSeconds?: number;
  notes?: string;
  orderIndex?: number;
}

export interface WorkoutInput {
  name: string;
  description?: string;
  difficulty: string;
  estimatedMinutes: number;
  category?: string;
  exercises?: WorkoutExerciseInput[];
}

export interface WorkoutUpdate {
  name?: string;
  description?: string;
  difficulty?: string;
  estimatedMinutes?: number;
  category?: string;
}

export interface WorkoutExercise {
  id: number;
  exerciseId: number;
  exerciseName?: string;
  muscleGroup?: string;
  equipment?: string;
  sets: number;
  reps: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  restSeconds?: number | null;
  /** @nullable */
  notes?: string | null;
  orderIndex: number;
}

export interface WorkoutDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  difficulty: string;
  estimatedMinutes: number;
  /** @nullable */
  category?: string | null;
  isAiGenerated?: boolean;
  createdAt: string;
  exercises: WorkoutExercise[];
}

export interface WorkoutLog {
  id: number;
  /** @nullable */
  workoutId?: number | null;
  workoutName: string;
  completedAt: string;
  durationMinutes: number;
  /** @nullable */
  totalVolume?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rating?: number | null;
}

export interface WorkoutLogSetInput {
  exerciseId: number;
  setNumber: number;
  reps: number;
  weight?: number;
  notes?: string;
}

export interface WorkoutLogInput {
  workoutId?: number;
  workoutName: string;
  completedAt: string;
  durationMinutes: number;
  notes?: string;
  rating?: number;
  sets?: WorkoutLogSetInput[];
}

export interface WorkoutLogSet {
  id: number;
  exerciseId: number;
  exerciseName?: string;
  setNumber: number;
  reps: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface WorkoutLogDetail {
  id: number;
  /** @nullable */
  workoutId?: number | null;
  workoutName: string;
  completedAt: string;
  durationMinutes: number;
  /** @nullable */
  totalVolume?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rating?: number | null;
  sets: WorkoutLogSet[];
}

export interface ProgressSummary {
  totalWorkouts: number;
  totalMinutes: number;
  currentStreak: number;
  longestStreak: number;
  totalVolume: number;
  workoutsThisWeek: number;
  workoutsThisMonth: number;
  /** @nullable */
  favoriteCategory?: string | null;
}

export interface WeeklyActivity {
  week: string;
  workoutCount: number;
  totalMinutes: number;
  totalVolume: number;
}

export interface PersonalRecord {
  exerciseId: number;
  exerciseName: string;
  muscleGroup?: string;
  /** @nullable */
  maxWeight?: number | null;
  maxReps: number;
  achievedAt: string;
}

export interface UserProfile {
  id: number;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weightGoal?: number | null;
  weightUnit: string;
  fitnessGoal: string;
  fitnessLevel: string;
  weeklyWorkoutTarget: number;
  /** @nullable */
  preferredEquipment?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  age?: number;
  weight?: number;
  height?: number;
  weightGoal?: number;
  weightUnit?: string;
  fitnessGoal?: string;
  fitnessLevel?: string;
  weeklyWorkoutTarget?: number;
  preferredEquipment?: string;
}

export interface WeightLogEntry {
  id: number;
  weight: number;
  unit: string;
  /** @nullable */
  notes?: string | null;
  loggedAt: string;
}

export interface WeightLogInput {
  weight: number;
  unit?: string;
  notes?: string;
}

export type ListExercisesParams = {
category?: string;
muscle?: string;
search?: string;
};

export type ListWorkoutLogsParams = {
limit?: number;
};