# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FitForge Workout App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"

  # ── Exercise Library ──────────────────────────────────────────────────────
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: muscle
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found
    delete:
      operationId: deleteExercise
      tags: [exercises]
      summary: Delete a custom exercise
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Workout Plans ─────────────────────────────────────────────────────────
...
  /workout-logs:
    get:
      operationId: listWorkoutLogs
      tags: [logs]
      summary: List completed workout sessions
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutLog"
    post:
      operationId: createWorkoutLog
      tags: [logs]
      summary: Log a completed workout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLog"
  /workout-logs/{id}:
    get:
      operationId: getWorkoutLog
      tags: [logs]
      summary: Get workout log details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLogDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteWorkoutLog
      tags: [logs]
      summary: Delete workout log
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Progress & Analytics ──────────────────────────────────────────────────
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile and fitness goals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profile/weight-log:
    get:
      operationId: listWeightLogs
      tags: [profile]
      summary: List weight log entries
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightLogEntry"
    post:
      operationId: createWeightLog
      tags: [profile]
      summary: Add a weight log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightLogEntry"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
    UpdateMeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string

    # ── Exercise ──────────────────────────────────────────────────────────
    Exercise:
      type: object
      required: [id, name, category, muscleGroup, equipment, isCustom, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        muscleGroup:
          type: string
        equipment:
          type: string
        instructions:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isCustom:
          type: boolean
        createdAt:
          type: string
    ExerciseInput:
      type: object
      required: [name, category, muscleGroup, equipment]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        muscleGroup:
          type: string
        equipment:
          type: string
        instructions:
          type: string

    # ── Workout Plan ──────────────────────────────────────────────────────
...
    WorkoutLog:
      type: object
      required: [id, workoutName, completedAt, durationMinutes]
      properties:
        id:
          type: integer
        workoutId:
          type: ["integer", "null"]
        workoutName:
          type: string
        completedAt:
          type: string
        durationMinutes:
          type: integer
        totalVolume:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
    WorkoutLogInput:
      type: object
      required: [workoutName, completedAt, durationMinutes]
      properties:
        workoutId:
          type: integer
        workoutName:
          type: string
        completedAt:
          type: string
        durationMinutes:
          type: integer
        notes:
          type: string
        rating:
          type: integer
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutLogSetInput"
    WorkoutLogSetInput:
      type: object
      required: [exerciseId, setNumber, reps]
      properties:
        exerciseId:
          type: integer
        setNumber:
          type: integer
        reps:
          type: integer
        weight:
          type: number
        notes:
          type: string
    WorkoutLogSet:
      type: object
      required: [id, exerciseId, setNumber, reps]
      properties:
        id:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        setNumber:
          type: integer
        reps:
          type: integer
        weight:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    WorkoutLogDetail:
      type: object
      required: [id, workoutName, completedAt, durationMinutes, sets]
      properties:
        id:
          type: integer
        workoutId:
          type: ["integer", "null"]
        workoutName:
          type: string
        completedAt:
          type: string
        durationMinutes:
          type: integer
        totalVolume:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutLogSet"

    # ── Progress & Analytics ──────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, name, weightUnit, fitnessGoal, fitnessLevel, weeklyWorkoutTarget, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: ["integer", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        weightGoal:
          type: ["number", "null"]
        weightUnit:
          type: string
        fitnessGoal:
          type: string
        fitnessLevel:
          type: string
        weeklyWorkoutTarget:
          type: integer
        preferredEquipment:
          type: ["string", "null"]
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        weight:
          type: number
        height:
          type: number
        weightGoal:
          type: number
        weightUnit:
          type: string
        fitnessGoal:
          type: string
        fitnessLevel:
          type: string
        weeklyWorkoutTarget:
          type: integer
        preferredEquipment:
          type: string
    WeightLogEntry:
      type: object
      required: [id, weight, unit, loggedAt]
      properties:
        id:
          type: integer
        weight:
          type: number
        unit:
          type: string
        notes:
          type: ["string", "null"]
        loggedAt:
          type: string
    WeightLogInput:
      type: object
      required: [weight]
      properties:
        weight:
          type: number
        unit:
          type: string
        notes:
          type: string