# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitForge Workout App API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  name?: string;
  age?: number;
  weight?: number;
  height?: number;
  weightGoal?: number;
  weightUnit?: string;
  fitnessGoal?: string;
  fitnessLevel?: string;
  weeklyWorkoutTarget?: number;
  preferredEquipment?: string;
}