# lib/api-zod/src/generated/types/weightLogEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitForge Workout App API
 * OpenAPI spec version: 0.1.0
 */

export interface WeightLogEntry {
  id: number;
  weight: number;
  unit: string;
  /** @nullable */
  notes?: string | null;
  loggedAt: string;
}