# lib/api-zod/src/generated/types/workoutLogInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitForge Workout App API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutLogSetInput } from './workoutLogSetInput';

export interface WorkoutLogInput {
  workoutId?: number;
  workoutName: string;
  completedAt: string;
  durationMinutes: number;
  notes?: string;
  rating?: number;
  sets?: WorkoutLogSetInput[];
}