# lib/api-zod/src/generated/types/workoutLogSet.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitForge Workout App API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkoutLogSet {
  id: number;
  exerciseId: number;
  exerciseName?: string;
  setNumber: number;
  reps: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  notes?: string | null;
}